Implement "add" for userstyle
Unfortunatly this commit is a bit big but I will try to break down the
main things:
 * Add CLI structure (clap.yml)
 * Add config creation (config.rs)
 * Add userstyle add functionality

The biggest changes in this commit are the config creation and the
userstyle add functionality. These interact with the user to create a
new config and load a style from userstyles.org, then getting all
settings and return the raw css.