Add add subcommand implementation
It is now possible to remove a style. Removing it from the config is
fairly straight forward. To be able to remove it from the
userContent.css without having to update all styles it had to be wrapped
in a CSS comment. Then everything from start comment to end comment can
be easily removed.