Add userChrome option
It is now possible to write to the userChrome instead of the userContent
file. There also have been some bugs with removing styles that now are
fixed. Based on some basic manual testing everything seems to work now.
The code is ready for a first "release" but it still needs documentation
on how to use it.