Fix file not found errors
If the `chrome` directory, `userContent.css` or `userChrome.css` do not
exist the program used to fail running `add` and `update`. Now `chrome`
is created whenever running `add` or `update` and it does not exist.
The `userChrome.css` and `userContent.css` missing do not cancel the
removing process anymore. It just removes it from the config because
there is no file to remove it from.