Add graceful error recovery
Adding and Removing styles now doesn't have the possibility for only
config or userChrome/-Content to be updated. Whenever writing the
userChrome/-Content fails, the config is recovered.

Due to update calling only remove and then add, without writing to
config or userChrome/-Content, this now also recovers gracefully.
However it is still possible that during an update a style is removed,
but readding it fails.

This fixes #1.