Recover from update failure
With this addition `rum update` will make sure to always exit in a valid
state. It should not be possible for it to break a config or target
file. The intended betavior is that when an operation fails, the config
and target file are reset to the state before updating.

If the recovery fails the program crashes with an error that config or
target might be corrupted. This should never actually happen and it
would probably indicate a bug, or a user really trying to break stuff.

This fixes #2.