Add option to disable a style temporarily
Using `rum update -t <STYLE>...` it is now possible to toggle the status
of a style. This makes it possible to disable a style without removing
it completely.

This works with the old configuration but it adds a new field to the
config. So any tool relying on the current config structure might need
to be changed.

A few minor changes include color not being reset after `eprintln`
resulting in normal text being red. Generally the output has been
tweaked a bit to make it easier to implement this feature.