Refactor main.rs
The first step of a bunch of refactoring changes to make testing easier.

This adds a global `READER` to the main.rs which can be used from
everywhere without having to pass around parameters. It still allows
mocking using a Cursor.

A `submatches` method has been added to cut down the main method's line
length a bit and prevent some useless repetition.

The `error!` macro has been added for printing to stderr with a red text
and then automatically resetting the text color back to white. This
hopefully just make the errors in the code a little more readable.